requirements.txt
djangorestframework==3.10.3
django==2.2.5